% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mopSplitFixedPattern}
\alias{mopSplitFixedPattern}
\title{mopSplitFixedPattern}
\usage{
mopSplitFixedPattern(s, pattern, splitLength = 2, cols = NULL)
}
\arguments{
\item{string}{}
}
\value{
string
}
\description{
mopSplitFixedPattern
}
\examples{
\dontrun{
}
}

