% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runMopper}
\alias{runMopper}
\title{runMopper}
\usage{
runMopper(dp, cleanId, opts, colIds = NULL, appendCol = FALSE)
}
\arguments{
\item{s}{string or dataframe}

\item{pos}{string, "first" or "last"}
}
\value{
string
}
\description{
runMopper
}
\examples{
\dontrun{
}
}

